event_machine: join reactor_thread if it is already running
...rather than falling through worker_loop

Prior to the application of this patch, if an EventMachine
reactor_thread has already been started elsewhere before the
worker_loop is entered, the worker_loop exits as a second call
to EM.run does not block the current thread.

This patch causes the worker_loop thread to join the
reactor_thread if it is running.

[ew: commit message formatting]

Signed-off-by: Eric Wong <normalperson@yhbt.net>