update dependencies for Ruby 2.2.0dev
This will allow me to test for unintentional breakage in 2.2.0.

Part of the reason for putting this project on maintenance mode
is because many of the libraries we depend on have not kept up
with the latest changes to Ruby.  So we will disable many tests
for 2.2+ to ensure the core parts remain working.