support rack 2.x
This requires updating to unicorn 5.1+ to relax the rack
requirement.  Unlike unicorn, Rainbows! is considerably
larger and cannot function without rack installed.

There were no code changes required to pass the test suite;
but some other optional gems needed to be updated for testing.

Support for async_sinatra is no longer tested as that is tied
to rack 1.x; but the remaining EventMachine-dependent parts
remain tested.

Thanks to Julia López Aladro for reminding us on the mailing list:
https://bogomips.org/rainbows-public/CAPhSs6OAU+motd+MF=WOSek9==StLjEQJsNwAO3-aRWCZzhs=g@mail.gmail.com/