doc: switch bogomips.org to yhbt.net
bogomips.org is due to expire, soon, and I'm not willing to pay
extortionists at Ethos Capital/PIR/ICANN to keep a .org.  So
it's at yhbt.net, for now...  Identity is overrated.

Tor users can use .onions and kick ICANN to the curb:

  torsocks w3m http://rainbows.ou63pmih66umazou.onion/
  torsocks git clone http://ou63pmih66umazou.onion/rainbows.git/
  torsocks w3m http://ou63pmih66umazou.onion/rainbows-public/

While we're at it, switch news.gmane.org => news.gmane.io
(but I suspect that'll need to be resynched since our mail
"List-Id:" header is changing).